+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Added 'sm' (thinner variant) for `Progress` component
+-   Handle size property on `Progress` component
 
 ## [1.0.9] - 2024-01-23